Implemented DevMode and some layout fixes

+++ src/services/apiOrders.js
@@ -1,4 +1,5 @@
 import { supabase } from "../supabase";
+import { DEV_MODE, DEV_MODE_MESSAGE } from "../shared/constansts";
 
 export async function getOrders() {
   const { data: orders, error } = await supabase
@@ -10,16 +11,12 @@ export async function getOrders() {
   return orders;
 }
 
-export async function addOrder(data) {
-  const { error } = await supabase.from("orders").insert(data);
-
-  if (error) throw error;
-}
-
 export async function deleteOrder(id) {
-  const { error } = await supabase.from("orders").delete().eq("id", id);
+  if (!DEV_MODE) {
+    const { error } = await supabase.from("orders").delete().eq("id", id);
 
-  if (error) throw error;
+    if (error) throw error;
+  } else throw new Error(DEV_MODE_MESSAGE);
 }
 
 export async function editOrderStatus(status, id) {
